Upgradeability check: - Fix bug: the json output was always enabled - Move output_json to a shared module
